+++ src/textual/app.py
@@ -97,8 +97,11 @@ from .widget import AwaitMount, Widget
 if TYPE_CHECKING:
     from typing_extensions import Coroutine, TypeAlias
 
+    # Unused & ignored imports are needed for the docs to link to these objects:
+    from .css.query import WrongType  # type: ignore  # noqa: F401
     from .devtools.client import DevtoolsClient
     from .pilot import Pilot
+    from .widget import MountError  # type: ignore  # noqa: F401
 
 PLATFORM = platform.system()
 WINDOWS = PLATFORM == "Windows"
@@ -137,6 +140,7 @@ ComposeResult = Iterable[Widget]
 RenderResult = RenderableType
 
 AutopilotCallbackType: TypeAlias = "Callable[[Pilot], Coroutine[Any, Any, None]]"
+"""Signature for valid callbacks that can be used to control apps."""
 
 
 class AppError(Exception):
@@ -167,6 +171,7 @@ CSSPathType = Union[
     PurePath,
     List[Union[str, PurePath]],
 ]
+"""Valid ways of specifying paths to CSS files."""
 
 CallThreadReturnType = TypeVar("CallThreadReturnType")
 
@@ -186,17 +191,7 @@ class _NullFile:
 
 @rich.repr.auto
 class App(Generic[ReturnType], DOMNode):
-    """The base class for Textual Applications.
-
-    Args:
-        driver_class: Driver class or `None` to auto-detect. This will be used by some Textual tools.
-        css_path: Path to CSS or `None` to use the `CSS_PATH` class variable.
-            To load multiple CSS files, pass a list of strings or paths which will be loaded in order.
-        watch_css: Reload CSS if the files changed. This is set automatically if you are using `textual run` with the `dev` switch.
-
-    Raises:
-        CssPathError: When the supplied CSS path(s) are an unexpected type.
-    """
+    """The base class for Textual Applications."""
 
     CSS: ClassVar[str] = ""
     """Inline CSS, useful for quick scripts. This is loaded after CSS_PATH,
@@ -218,8 +213,10 @@ class App(Generic[ReturnType], DOMNode):
     """
 
     SCREENS: ClassVar[dict[str, Screen | Callable[[], Screen]]] = {}
+    """Screens associated with the app for the lifetime of the app."""
     _BASE_PATH: str | None = None
     CSS_PATH: ClassVar[CSSPathType | None] = None
+    """File paths to load CSS from."""
 
     TITLE: str | None = None
     """A class variable to set the *default* title for the application.
@@ -255,6 +252,20 @@ class App(Generic[ReturnType], DOMNode):
         css_path: CSSPathType | None = None,
         watch_css: bool = False,
     ):
+        """Create an instance of an app.
+
+        Args:
+            driver_class: Driver class or `None` to auto-detect.
+                This will be used by some Textual tools.
+            css_path: Path to CSS or `None` to use the `CSS_PATH` class variable.
+                To load multiple CSS files, pass a list of strings or paths which
+                will be loaded in order.
+            watch_css: Reload CSS if the files changed. This is set automatically if
+                you are using `textual run` with the `dev` switch.
+
+        Raises:
+            CssPathError: When the supplied CSS path(s) are an unexpected type.
+        """
         super().__init__()
         self.features: frozenset[FeatureFlag] = parse_features(os.getenv("TEXTUAL", ""))
 
@@ -406,7 +417,7 @@ class App(Generic[ReturnType], DOMNode):
 
     @property
     def return_value(self) -> ReturnType | None:
-        """The return value of the app, or `None` if it as not yet been set.
+        """The return value of the app, or `None` if it has not yet been set.
 
         The return value is set when calling [exit][textual.app.App.exit].
         """
@@ -414,10 +425,11 @@ class App(Generic[ReturnType], DOMNode):
 
     @property
     def children(self) -> Sequence["Widget"]:
-        """A view on to the App's children.
+        """A view onto the app's immediate children.
 
         This attribute exists on all widgets.
-        In the case of the App, it will only every contain a single child, which will be the currently active screen.
+        In the case of the App, it will only ever contain a single child, which will
+        be the currently active screen.
 
         Returns:
             A sequence of widgets.
@@ -499,7 +511,7 @@ class App(Generic[ReturnType], DOMNode):
 
     @property
     def screen_stack(self) -> Sequence[Screen]:
-        """The current screen stack.
+        """A snapshot of the current screen stack.
 
         Returns:
             A snapshot of the current state of the screen stack.
@@ -523,7 +535,7 @@ class App(Generic[ReturnType], DOMNode):
 
     @property
     def focused(self) -> Widget | None:
-        """The widget that is focused on the currently active screen.
+        """The widget that is focused on the currently active screen, or `None`.
 
         Focused widgets receive keyboard input.
 
@@ -534,7 +546,7 @@ class App(Generic[ReturnType], DOMNode):
 
     @property
     def namespace_bindings(self) -> dict[str, tuple[DOMNode, Binding]]:
-        """Get current bindings.
+        """Get currently active bindings.
 
         If no widget is focused, then app-level bindings are returned.
         If a widget is focused, then any bindings present in the active screen and app are merged and returned.
@@ -542,8 +554,7 @@ class App(Generic[ReturnType], DOMNode):
         This property may be used to inspect current bindings.
 
         Returns:
-
-            A mapping of keys on to node + binding.
+            A mapping of keys onto pairs of nodes and bindings.
         """
 
         namespace_binding_map: dict[str, tuple[DOMNode, Binding]] = {}
@@ -650,7 +661,7 @@ class App(Generic[ReturnType], DOMNode):
 
     @property
     def screen(self) -> Screen:
-        """Screen: The current screen.
+        """The current active screen.
 
         Returns:
             The currently active (visible) screen.
@@ -689,7 +700,7 @@ class App(Generic[ReturnType], DOMNode):
 
     @property
     def log(self) -> Logger:
-        """Textual log interface.
+        """The textual logger.
 
         Example:
             ```python
@@ -763,12 +774,16 @@ class App(Generic[ReturnType], DOMNode):
         *args,
         **kwargs,
     ) -> CallThreadReturnType:
-        """Run a callback from another thread.
+        """Run a callable from another thread, and return the result.
 
         Like asyncio apps in general, Textual apps are not thread-safe. If you call methods
         or set attributes on Textual objects from a thread, you may get unpredictable results.
 
         This method will ensure that your code runs within the correct context.
+
+        !!! tip
+
+            Consider using [post_message][textual.message_pump.MessagePump.post_message] which is also thread-safe.
 
         Args:
             callback: A callable to run.
@@ -778,6 +793,9 @@ class App(Generic[ReturnType], DOMNode):
         Raises:
             RuntimeError: If the app isn't running or if this method is called from the same
                 thread where the app is running.
+
+        Returns:
+            The result of the callback.
         """
 
         if self._loop is None:
@@ -1162,14 +1180,15 @@ class App(Generic[ReturnType], DOMNode):
 
         Args:
             id: The ID of the node to search for.
-            expect_type: Require the object be of the supplied type, or None for any type.
+            expect_type: Require the object be of the supplied type,
+                or use `None` to apply no type restriction.
 
         Returns:
             The first child of this node with the specified ID.
 
         Raises:
-            NoMatches: if no children could be found for this ID
-            WrongType: if the wrong type was found.
+            NoMatches: If no children could be found for this ID.
+            WrongType: If the wrong type was found.
         """
         return (
             self.screen.get_child_by_id(id)
@@ -1462,7 +1481,6 @@ class App(Generic[ReturnType], DOMNode):
         Note that uninstalling a screen is only required if you have previously installed it
         with [install_screen][textual.app.App.install_screen].
         Textual will also uninstall screens automatically on exit.
-
 
         Args:
             screen: The screen to uninstall or the name of a installed screen.
@@ -1836,6 +1854,7 @@ class App(Generic[ReturnType], DOMNode):
             *widgets: The widget(s) to register.
             before: A location to mount before.
             after: A location to mount after.
+
         Returns:
             List of modified widgets.
         """
@@ -2140,7 +2159,7 @@ class App(Generic[ReturnType], DOMNode):
                 or None to use app.
 
         Returns:
-            True if the event has handled.
+            True if the event has been handled.
         """
         if isinstance(action, str):
             target, params = actions.parse(action)