+++ src/textual/errors.py
@@ -14,4 +14,8 @@ class RenderError(TextualError):
 
 
 class DuplicateKeyHandlers(TextualError):
-    """More than one handler for a single key press. E.g. key_ctrl_i and key_tab handlers both found on one object."""
+    """More than one handler for a single key press.
+
+    For example, if the handlers `key_ctrl_i` and `key_tab` were defined on the same
+    widget, then this error would be raised.
+    """